Rename event, move to renderer package
