+++ fabric-events-lifecycle-v0/src/main/resources/fabric-events-lifecycle-v0.mixins.json
@@ -7,8 +7,7 @@
     "MixinWorld"
   ],
   "client": [
-    "MixinMinecraftClient",
-    "MixinWorldRenderer"
+    "MixinMinecraftClient"
   ],
   "injectors": {
     "defaultRequire": 1
