+++ fabric-rendering-v0/src/main/java/net/fabricmc/fabric/mixin/client/render/MixinWorldRenderer.java
@@ -14,20 +14,20 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.mixin.eventslifecycle;
+package net.fabricmc.fabric.mixin.client.render;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import net.fabricmc.fabric.api.event.client.InvalidateRenderStateCallback;
+import net.fabricmc.fabric.api.client.render.InvalidateRenderStateCallback;
 import net.minecraft.client.render.WorldRenderer;
 
 @Mixin(WorldRenderer.class)
 public abstract class MixinWorldRenderer {
     @Inject(method = "reload", at = @At("HEAD"))
     private void onReload(CallbackInfo ci) {
-        InvalidateRenderStateCallback.EVENT.invoker().onReload();
+        InvalidateRenderStateCallback.EVENT.invoker().onInvalidate();
     }
 }
