+++ fabric-rendering-v0/src/main/resources/fabric-rendering-v0.mixins.json
@@ -6,7 +6,8 @@
     "MixinBlockColorMap",
     "MixinBlockEntityRenderManager",
     "MixinEntityRenderManager",
-    "MixinItemColorMap"
+    "MixinItemColorMap",
+    "MixinWorldRenderer"
   ],
   "injectors": {
     "defaultRequire": 1
